requirements.txt
flask~=1.1.2
werkzeug~=1.0.1
sqlalchemy~=1.3.20

progressbar2
python-dateutil~=2.8.1
pandas~=1.1.4
passlib~=1.7.4
alembic~=1.4.3